+++ src/main/java/com/github/cc3002/citricjuice/model/board/DropPanel.java
@@ -1,6 +1,6 @@
 package com.github.cc3002.citricjuice.model.board;
 
-import com.github.cc3002.citricjuice.model.PlayerUnit;
+import com.github.cc3002.citricjuice.model.unit.PlayerUnit;
 import org.jetbrains.annotations.NotNull;
 
 public class DropPanel extends AbstractPanel {
