+++ src/main/java/com/github/cc3002/citricjuice/model/unit/IUnit.java
@@ -1,4 +1,4 @@
-package com.github.cc3002.citricjuice.model;
+package com.github.cc3002.citricjuice.model.unit;
 
 /**
  * interface for the units
