+++ src/main/java/com/github/cc3002/citricjuice/model/unit/PlayerUnit.java
@@ -1,4 +1,4 @@
-package com.github.cc3002.citricjuice.model;
+package com.github.cc3002.citricjuice.model.unit;
 
 /**
  * This class represents a playerUnit in the game 99.7% Citric Liquid.
