+++ src/test/java/com/github/cc3002/citricjuice/model/board/PanelTest.java
@@ -1,6 +1,6 @@
 package com.github.cc3002.citricjuice.model.board;
 
-import com.github.cc3002.citricjuice.model.PlayerUnit;
+import com.github.cc3002.citricjuice.model.unit.PlayerUnit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
